powerpc/mm/hash64: Update comment about size of the vmalloc region
The comment about the size of the vmalloc region is very out of date.
It talks about Book3E, but we split Book3E out of this code a while
ago. And we also changed the size calculation too, it's no longer
half. So update the comment to reflect the current behaviour.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>