powerpc/mm/hash64: Document the size in bytes of H_Pxx_INDEX_SIZE
Add comments describing the size in bytes of the various levels of the
page table tree, to make it clear what the sizes are without having to
also look up other definitions. The code uses sizeof() so in theory
these could skew vs the code, but the size of pgd_t etc. is unlikely
to change very often.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>